api: Split server.go to  *_service.go
